Move unix socket client to pkg/util
